refactor: thematic merges + state/runtime grouping + paneHeader → components
Three coherent regroupings on top of the Phase 1 reorg:

- utils/commitUrl + utils/commitMetrics → utils/commit.ts (single
  commit-related helpers module). Tests merged into utils/commit.test.ts.
- localFlag + sourceRecents move into state/runtime/ — sourceRecents
  is parallel to sourceContext (both are source-related runtime
  state); localFlag holds runtime UI flags too. state/runtime/ now owns
  ALL runtime state, persisted or not.
- views/shell/paneHeader → views/components/paneHeader. paneHeader is a
  reusable element used by every pane — it belongs in components/, not
  in shell/ (which is reserved for page-level chrome: header, footer,
  sidebars).

The "two persistence systems" smell (localFlag + sourceRecents do
direct localStorage writes outside the persist.ts pipeline) is
deferred to its own follow-up phase — that's an architectural change,
not a file move.
