+++ app/src/state/runtime/sourceRecents.ts
@@ -1,5 +1,5 @@
-// utils/sourceRecents.ts — Tiny localStorage-backed list of recently-opened
-// sources. Used by the source picker modal to show one-click reload rows.
+// state/runtime/sourceRecents.ts — Tiny localStorage-backed list of recently-
+// opened sources. Used by the source picker modal to show one-click reload rows.
 
 const KEY = 'codecity:recents';
 const MAX = 10;
