+++ app/src/views/panes/controlsPane.ts
@@ -76,7 +76,7 @@ import {
 import { KEY_BINDINGS } from '@/constants';
 import { FadeDetail } from '@/types';
 import { makeLucideIcon } from '@/views/components/icon';
-import { buildPaneHeader } from '@/views/shell/paneHeader';
+import { buildPaneHeader } from '@/views/components/paneHeader';
 
 // Structural store shape used by all the widget builders. Covers nanostores
 // `map<T>()` (with .setKey) and falls back to .set for atom-like stores.
